Move symmetric cipher decls to own header
