+++ src/lib/symmetric.h
@@ -1,0 +1,136 @@
+/*
+ * Copyright (c) 2017, [Ribose Inc](https://www.ribose.com).
+ * Copyright (c) 2009 The NetBSD Foundation, Inc.
+ * All rights reserved.
+ *
+ * This code is originally derived from software contributed to
+ * The NetBSD Foundation by Alistair Crooks (agc@netbsd.org), and
+ * carried further by Ribose Inc (https://www.ribose.com).
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE NETBSD FOUNDATION, INC. AND CONTRIBUTORS
+ * ``AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+ * TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE FOUNDATION OR CONTRIBUTORS
+ * BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
+/*
+ * Copyright (c) 2005-2008 Nominet UK (www.nic.uk)
+ * All rights reserved.
+ * Contributors: Ben Laurie, Rachel Willmer. The Contributors have asserted
+ * their moral rights under the UK Copyright Design and Patents Act 1988 to
+ * be recorded as the authors of this copyright work.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License.
+ *
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef SYMMETRIC_CRYPTO_H_
+#define SYMMETRIC_CRYPTO_H_
+
+/** pgp_crypt_t */
+struct pgp_crypt_t {
+	pgp_symm_alg_t	alg;
+	size_t			blocksize;
+	size_t			keysize;
+	void 			(*set_iv)(pgp_crypt_t *, const uint8_t *);
+	void			(*set_crypt_key)(pgp_crypt_t *, const uint8_t *);
+	int			(*base_init)(pgp_crypt_t *);
+	void			(*decrypt_resync)(pgp_crypt_t *);
+	/* encrypt/decrypt one block */
+	void			(*block_encrypt)(pgp_crypt_t *, uint8_t *, const uint8_t *);
+	void			(*block_decrypt)(pgp_crypt_t *, uint8_t *, const uint8_t *);
+	/* Standard CFB encrypt/decrypt (as used by Sym Enc Int Prot packets) */
+	void 			(*cfb_encrypt)(pgp_crypt_t *, uint8_t *, const uint8_t *, size_t);
+	void			(*cfb_decrypt)(pgp_crypt_t *, uint8_t *, const uint8_t *, size_t);
+	void			(*decrypt_finish)(pgp_crypt_t *);
+	uint8_t			iv[PGP_MAX_BLOCK_SIZE];
+	uint8_t			civ[PGP_MAX_BLOCK_SIZE];
+	uint8_t			siv[PGP_MAX_BLOCK_SIZE];
+		/* siv is needed for weird v3 resync */
+	uint8_t			key[PGP_MAX_KEY_SIZE];
+	int			num;
+		/* num is offset for CFB */
+	struct botan_block_cipher_struct			*block_cipher_obj;
+
+};
+
+pgp_symm_alg_t pgp_str_to_cipher(const char *name);
+unsigned pgp_block_size(pgp_symm_alg_t);
+unsigned pgp_key_size(pgp_symm_alg_t);
+unsigned pgp_is_sa_supported(pgp_symm_alg_t);
+
+size_t pgp_decrypt_se(pgp_crypt_t *, void *, const void *, size_t);
+size_t pgp_encrypt_se(pgp_crypt_t *, void *, const void *, size_t);
+size_t pgp_decrypt_se_ip(pgp_crypt_t *, void *, const void *, size_t);
+size_t pgp_encrypt_se_ip(pgp_crypt_t *, void *, const void *, size_t);
+unsigned pgp_is_sa_supported(pgp_symm_alg_t);
+
+int pgp_crypt_any(pgp_crypt_t *, pgp_symm_alg_t);
+int pgp_decrypt_init(pgp_crypt_t *);
+int pgp_encrypt_init(pgp_crypt_t *);
+
+size_t pgp_decrypt_se(pgp_crypt_t *, void *, const void *, size_t);
+size_t pgp_encrypt_se(pgp_crypt_t *, void *, const void *, size_t);
+size_t pgp_decrypt_se_ip(pgp_crypt_t *, void *, const void *, size_t);
+size_t pgp_encrypt_se_ip(pgp_crypt_t *, void *, const void *, size_t);
+
+
+
+#if 0
+// Initialize a cipher object
+void pgp_cipher_create(pgp_cipher_t** cipher, pgp_symm_alg_t algo);
+
+// Destroy a cipher object, freeing all resources
+void pgp_cipher_finish(pgp_cipher_t* cipher);
+
+// Return blocksize of cipher or 0 if error
+size_t pgp_cipher_blocksize(const pgp_cipher_t* cipher);
+
+// Return key length of cipher or 0 if error
+size_t pgp_cipher_keysize(const pgp_cipher_t* cipher);
+
+// IV is always pgp_cipher_blocksize(cipher) bytes long
+void pgp_cipher_set_iv(pgp_cipher_t* cipher, const uint8_t iv[]);
+
+// Key is always pgp_cipher_keysize(cipher) bytes long
+void pgp_cipher_set_key(pgp_cipher_t* cipher, const uint8_t key[]);
+
+void pgp_cipher_decrypt_resync(pgp_cipher_t* cipher);
+
+void pgp_cipher_block_encrypt(const pgp_cipher_t* cipher, uint8_t *block_out, const uint8_t *block_in);
+void pgp_cipher_block_decrypt(const pgp_cipher_t* cipher, uint8_t *block_out, const uint8_t *block_in);
+
+void pgp_cipher_cfb_encrypt(const pgp_cipher_t* cipher, uint8_t *text_out, const uint8_t *text_in, size_t len);
+void pgp_cipher_cfb_decrypt(const pgp_cipher_t* cipher, uint8_t *text_out, const uint8_t *text_in, size_t len);
+
+// FIXME remove this
+#define pgp_crypt_t pgp_cipher_t
+#endif
+
+
+#endif
